+++ src/pathtracer/samplebuffer.hpp
@@ -12,21 +12,21 @@
 class SampleBuffer
 {
   public:
-    SampleBuffer(size_t w, size_t h)
+    SampleBuffer(unsigned int w, unsigned int h)
       : m_width(w), m_height(h), m_samples(0), m_buffer(w * h)
     {
       assert(w > 0 && h > 0);
     }
 
-    size_t width() const { return m_width; }
-    size_t height() const { return m_height; }
+    unsigned int width() const { return m_width; }
+    unsigned int height() const { return m_height; }
 
-    size_t size() const
+    unsigned int size() const
     {
       return m_buffer.size();
     }
 
-    size_t samples() const
+    unsigned int samples() const
     {
       return m_samples;
     }
@@ -36,12 +36,12 @@ class SampleBuffer
       ++m_samples;
     }
 
-    const glm::vec3& at(size_t x, size_t y) const
+    const glm::vec3& at(unsigned int x, unsigned int y) const
     {
       return m_buffer[y * m_width + x];
     }
 
-    void add(size_t x, size_t y, const glm::vec3& v)
+    void add(unsigned int x, unsigned int y, const glm::vec3& v)
     {
       m_buffer[y * m_width + x] += v;
     }
@@ -66,8 +66,8 @@ class SampleBuffer
     }
 
   private:
-    size_t m_width, m_height;
-    size_t m_samples;
+    unsigned int m_width, m_height;
+    unsigned int m_samples;
 
     std::vector<glm::vec3> m_buffer;
 };
